requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
dj-database-url==1.2.0
Django==4.1.5
django-decouple==2.1
django-meta==2.1.0
django-on-heroku==1.1.2
django-session-timeout==0.1.0
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
whitenoise==6.3.0
